requirements.txt
loguru==0.7.2
cryptography==42.0.4
ccxt==4.1.65
pandas==2.0.3
openpyxl==3.1.2
questionary-2.0.1